frontend/requirements.txt
gradio
numpy
pillow
requests